+++ scrapy_zyte_api/providers.py
@@ -22,12 +22,17 @@ from zyte_common_items import (
     AutoArticleListPage,
     AutoArticleNavigationPage,
     AutoArticlePage,
+    AutoJobPostingNavigationPage,
     AutoJobPostingPage,
     AutoProductListPage,
     AutoProductNavigationPage,
     AutoProductPage,
+    CustomAttributes,
+    CustomAttributesMetadata,
+    CustomAttributesValues,
     Item,
     JobPosting,
+    JobPostingNavigation,
     Product,
     ProductList,
     ProductNavigation,
@@ -35,7 +40,7 @@ from zyte_common_items import (
 from zyte_common_items.fields import is_auto_field
 
 from scrapy_zyte_api import Actions, ExtractFrom, Geolocation, Screenshot
-from scrapy_zyte_api._annotations import _ActionResult
+from scrapy_zyte_api._annotations import _ActionResult, _from_hashable
 from scrapy_zyte_api.responses import ZyteAPITextResponse
 
 try:
@@ -53,12 +58,14 @@ _ITEM_KEYWORDS: Dict[type, str] = {
     ArticleList: "articleList",
     ArticleNavigation: "articleNavigation",
     JobPosting: "jobPosting",
+    JobPostingNavigation: "jobPostingNavigation",
 }
 _AUTO_PAGES: Set[type] = {
     AutoArticlePage,
     AutoArticleListPage,
     AutoArticleNavigationPage,
     AutoJobPostingPage,
+    AutoJobPostingNavigationPage,
     AutoProductPage,
     AutoProductListPage,
     AutoProductNavigationPage,
@@ -76,8 +83,11 @@ class ZyteApiProvider(PageObjectInputProvider):
         ArticleNavigation,
         BrowserHtml,
         BrowserResponse,
+        CustomAttributes,
+        CustomAttributesValues,
         Geolocation,
         JobPosting,
+        JobPostingNavigation,
         Product,
         ProductList,
         ProductNavigation,
@@ -175,15 +185,14 @@ class ZyteApiProvider(PageObjectInputProvider):
                     )
                 zyte_api_meta["actions"] = []
                 for action in cls.__metadata__[0]:  # type: ignore[attr-defined]
-                    zyte_api_meta["actions"].append(
-                        {
-                            k: (
-                                dict(v)
-                                if isinstance(v, frozenset)
-                                else list(v) if isinstance(v, tuple) else v
-                            )
-                            for k, v in action
-                        }
+                    zyte_api_meta["actions"].append(_from_hashable(action))
+                continue
+            if cls_stripped in {CustomAttributes, CustomAttributesValues}:
+                custom_attrs_input, custom_attrs_options = cls.__metadata__[0]  # type: ignore[attr-defined]
+                zyte_api_meta["customAttributes"] = _from_hashable(custom_attrs_input)
+                if custom_attrs_options:
+                    zyte_api_meta["customAttributesOptions"] = _from_hashable(
+                        custom_attrs_options
                     )
                 continue
             kw = _ITEM_KEYWORDS.get(cls_stripped)
@@ -320,6 +329,27 @@ class ZyteApiProvider(PageObjectInputProvider):
                 else:
                     actions_result = None
                 result = AnnotatedInstance(Actions(actions_result), cls.__metadata__)  # type: ignore[attr-defined]
+                results.append(result)
+                continue
+            if cls_stripped is CustomAttributes and is_typing_annotated(cls):
+                custom_attrs_result = api_response.raw_api_response["customAttributes"]
+                result = AnnotatedInstance(
+                    CustomAttributes(
+                        CustomAttributesValues(custom_attrs_result["values"]),
+                        CustomAttributesMetadata.from_dict(
+                            custom_attrs_result["metadata"]
+                        ),
+                    ),
+                    cls.__metadata__,  # type: ignore[attr-defined]
+                )
+                results.append(result)
+                continue
+            if cls_stripped is CustomAttributesValues and is_typing_annotated(cls):
+                custom_attrs_result = api_response.raw_api_response["customAttributes"]
+                result = AnnotatedInstance(
+                    CustomAttributesValues(custom_attrs_result["values"]),
+                    cls.__metadata__,  # type: ignore[attr-defined]
+                )
                 results.append(result)
                 continue
             kw = _ITEM_KEYWORDS.get(cls_stripped)